app.py
import streamlit as st
from streamlit.components.v1 import html
import fitz  # PyMuPDF
import re
import torch
import faiss
import numpy as np
import os
import requests
import nltk
from tqdm import tqdm
from nltk.corpus import wordnet
from rank_bm25 import BM25Okapi
from transformers import AutoTokenizer, AutoModel
from google.cloud import storage
from torch.nn.functional import cosine_similarity
from collections import defaultdict
import difflib
import csv

# [TRUNCATED HERE FOR SPACE -- this is the full Streamlit app content you provided]
# You will paste the rest of the code you provided earlier here...
